requirements.txt
Flask
Flasgger
Flask-CORS
Keras
numpy
Pillow
pandas
requests
virtualenv
